✨ Added OrderItem related classes and updated AuthService
This commit adds the OrderItem entity, repository, service, and controller. This allows for managing (add, read, update, delete) the items within an order. Also it introduces the OrderItemDto for separation of concerns. The AuthService has been updated to add verification of user's association with an order. A Product model and its repository are also created.