:bug: Implement NotFoundException in various services
Implemented NotFoundException in order, order-item, payment, auth, and product services. This exception will be thrown when certain entities like order, product, order item or payment are not found in the database during operations such as fetch, update or delete. This allows for more precise error messaging and proper exception handling.