:sparkles: Introduced PagedResponse DTO for data pagination
A new data transfer object (DTO), PagedResponse, has been added to handle pagination data. This DTO has replaced the simple Iterable<ProductDto> return type in the getAll method of the ProductController and getProducts method in the ProductService. Furthermore, the ProductService's getProducts method was updated to create and return a PagedResponse object.