:recycle: Handle stock quantity during order item operations
The code now includes methods to handle the stock quantity of a product during order item operations. Alongside creating, updating, and deleting an order item, the corresponding quantity of the product in the stock will be adjusted accordingly. Also, default order status has been set to 'CREATED' and it has now the access mode set to 'READ_ONLY' in OrderDto.