🔧 Change default ports for services and UI
The commit updates the default ports used in the application. The changes include updating port numbers in docker-compose.yml, nginx.conf.template and .env files, also reflects the updates in the Dockerfile. This change allow services to run on different ports avoiding port conflicts and enhance the security of the application.