🔥 Remove entrypoint.sh, update API URLs and docker-compose
The entrypoint.sh script was removed from the froome-ui project. Additionally, URLs in the .env file have been updated to point to 'localhost' instead of the specific services. Modifications were carried out in the Dockerfile and docker-compose file. Also, a 'serve' script was added to package.json to run the 'vite preview'.