🎁 Add config-service and update services with cloud config
Config-service was added to serve as a central location for configuration, creating this as a new Spring Boot application. Modified the existing services order-service, payment-service, product-service and user-service to use configurations from the newly added config-service. Modified the docker-compose file to include config-service and the respective environment variables. Moved application.yml files to config-service and renamed accordingly.