🔧 Integrated config-service into all microservices
The existing microservices (product-service, payment-service, user-service, order-service) have been modified to import the configuration from config-service. This is evident from the changes in bootstrap.yml files which now includes the 'config' setting for each service.

Removal of 'spring.cloud.config.enabled' from the '*.yml' files in the config-service indicates these services are no longer using local configuration but are now hooked up with the config-service. Also, a new Dockerfile for config-service was added and the existing Dockerfiles of each microservice were updated to include the necessary dependencies for config-service lining up for a successful Docker Build.

In addition,  pom.xml files for each microservice and parent project were adjusted to include config-service as a module, and the config-service's pom.xml file was populated with proper plugin configurations.